♻️ 코드 수정: LoginLeft 로그인 시 타이머 기능 수정
