+++ src/components/login/LoginLeft.vue
@@ -15,6 +15,7 @@ import { useLoginState } from '@/stores/loginState';
 import { ref, onMounted } from 'vue';
 import axios from 'axios';
 import { useRouter } from 'vue-router';
+import { startTimer } from '@/utils/timer';
 
 const loginState = useLoginState();
 const router = useRouter();
@@ -57,8 +58,16 @@ const loginUser = async () => {
     }
 
     if (exp) {
-      const expirationDate = new Date(exp);
-      startTimer(expirationDate);
+      const expirationTime = new Date(`1970-01-01T${exp}Z`);
+      startTimer(expirationTime, (remaining) => {
+        if (remaining === '만료됨') {
+          alert('세션이 만료되었습니다. 다시 로그인하세요.');
+          loginState.resetState();
+          router.replace('/login');
+        } else {
+          console.log(`남은 시간: ${remaining}`);
+        }
+      });
     }
 
     alert(`${name}님, 환영합니다.`);
@@ -72,7 +81,6 @@ const loginUser = async () => {
 onMounted(async () => {
   checkLoginStatus();
 });
-
 </script>
   
   <style scoped>
